requirements-gpu.txt
jupyter
notebook
numpy
matplotlib
tensorflow-gpu
tqdm
plotly
pandas
networkx
scipy